python/demo-server/requirements.txt
fastapi
uvicorn
azure-identity
azure-mgmt-resource
azure-mgmt-containerservice
openai
python-dotenv
